Moved scene viewer main from mod.rs to main.rs per Rust conventions
